+++ components/measures/MeasuresTable.vue
@@ -18,15 +18,53 @@
 
       <b-table
         id="sensor-table"
+        selectable
+        :sticky-header="true"
         :busy.sync="isBusy"
         :items="measures"
         :fields="fields"
         :per-page="perPage"
         :current-page="currentPage"
-        small :dark="colormode == 'dark'"
+        small
+        :dark="colormode == 'dark'"
       >
+        <template #cell(actions)="row">
+          <a v-b-modal.confirmDestroy 
+          title="delete" 
+          @click="current_id = row.item.id"> <b-icon class="text-danger" icon="trash-fill"></b-icon></a>
+         <!-- <b-button
+            size="sm"
+            @click="info(row.item, row.index, $event.target)"
+            class="mr-1"
+          >
+           
+          </b-button> -->
+          <a class="float-right" href="#" @click.prevent="row.toggleDetails">
+            <!--{{ row.detailsShowing ? 'Hide' : <b-icon v-if="row.detailsShowing" class="text-danger" icon="caret-down">` }} Details -->
+              <b-icon v-if="row.detailsShowing"  icon="caret-up"/>
+              <b-icon v-if="!row.detailsShowing"  icon="caret-down"/>
+          </a>
+        </template>
+        <template #row-details="row">
+          <b-card>
+            <ul class="list-unstyled">
+              <li v-for="(value, key) in row.item" :key="key">
+                <strong>{{ key }}:</strong> {{ value }}
+              </li>
+            </ul>
+          </b-card>
+        </template>
       </b-table>
+      <b-pagination
+        v-model="currentPage"
+        :total-rows="rows"
+        :per-page="perPage"
+        aria-controls="sensor-table"
+      ></b-pagination>
     </b-overlay>
+    <b-modal id="confirmDestroy" title="Confirm deletation " @ok="destroy">
+      Do you confirm irrevocable deletation of this measure ?
+    </b-modal>
   </div>
 </template>
 <script>
@@ -38,6 +76,7 @@ export default {
   data() {
     return {
       measures: [],
+      current_id:'',
       show: false,
       isBusy: false,
       fields: [
@@ -46,10 +85,18 @@ export default {
         { key: 'measure_unit' },
         { key: 'measure_type' },
         { key: 'origin' },
-        { key: 'created_at', sortable: true },
+        {
+          key: 'created_at',
+          sortable: true,
+          formatter: (value, key, item) => {
+            return this.$moment(value).format('DD/MM/YYYY H:mm:s')
+          },
+        },
+        { key: 'actions', label: 'Actions' },
       ],
-      perPage: 6,
+      perPage: 20,
       currentPage: 1,
+      pageOptions: [5, 10, 15, 20, { value: 100, text: 'Show a lot' }],
     }
   },
   methods: {
@@ -61,6 +108,17 @@ export default {
         this.show = false
         this.isBusy = false
       })
+    },
+    async destroy() {
+      this.isBusy = true
+      await this.$axios
+        .delete(`/v1/measure/destroy/${this.current_id}`)
+        .then((response) => {
+          this.measures = this.measures.filter((p) => p.id !== this.current_id)
+          this.show = false
+          this.isBusy = false
+          this.$toast.success('Measure deleted',{ duration:2500, })
+        })
     },
   },
   mounted() {
@@ -83,5 +141,4 @@ export default {
   },
 }
 </script>
-<style>
-</style>
+<style></style>